Imp splitting of test_preprocessing.py
